requirements.txt
Django==5.0.4
djangorestframework==3.15.1
channels==4.1.0
daphne==4.1.0
PyJWT==2.8.0